Validações parte 1 - Controle de fluxos

+++ app/Http/Controllers/MarcaController.php
@@ -21,6 +21,7 @@ class MarcaController extends Controller
         //$marcas = Marca::all();
 
         $marcas = $this->marca->all();
+
         return $marcas;
     }
 
@@ -41,6 +42,11 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
+
         return $marca;
     }
 
@@ -52,7 +58,16 @@ class MarcaController extends Controller
         //$marca->update($request->all());
 
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return [
+            'erro' => 
+            'Impossível realizar a atualização. O recurso solicitado não existe'
+            ];
+        }
+
         $marca->update($request->all());
+
         return $marca;
     }
 
@@ -64,6 +79,14 @@ class MarcaController extends Controller
         //$marca->delete();
 
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return [
+            'erro' => 
+            'Impossível realizar a exclusão. O recurso solicitado não existe'
+            ];
+        }
+
         $marca->delete();
 
         return ['msg' => 'A marca foi removida com sucesso'];
